Move plan state and procedure to specific files in a plan module
