+++ syft/messaging/plan/__init__.py
@@ -1,0 +1,3 @@
+from syft.messaging.plan.plan import func2plan
+from syft.messaging.plan.plan import method2plan
+from syft.messaging.plan.plan import Plan
