+++ syft/serde/serde.py
@@ -55,8 +55,8 @@ from syft.generic.pointers.pointer_plan import PointerPlan
 from syft.generic.pointers.pointer_protocol import PointerProtocol
 from syft.generic.pointers.object_wrapper import ObjectWrapper
 from syft.messaging.plan import Plan
-from syft.messaging.plan import State
-from syft.messaging.plan import Procedure
+from syft.messaging.plan.state import State
+from syft.messaging.plan.procedure import Procedure
 from syft.messaging.protocol import Protocol
 from syft.messaging.message import Message
 from syft.messaging.message import Operation
